Create four_words_2.csv
The is a edited version from a copy of four_words.csv as the initial
file contained two spelling errors ("trumpet" and "krasser af").
In the first case the word was English and is now "trompet". In the second
case, the word "krasser" was not recorded in dictionary but apparently used
by some authors. The standard spelling is "kradser".